refactor: split deadline/time logic into deadline.go, timestr.go, slope.go
The "when is this goal due, and how much is needed" concept was fragmented
across beeminder.go, main.go, review.go, and utils.go — each holding half
of what the others needed. Loading a deadline calculation meant bouncing
between three files; new commands would inevitably duplicate the time-
string parsing that already existed.

Splits the fragmentation into three deep modules:

- deadline.go — losedate predicates (today/tomorrow/within) + relative/
  absolute formatters + ParseDuration. Also absorbs the main.go filter
  wrappers (isDueTodayFilter, isDueTomorrowFilterAt, dueLaterTodayAt)
  that combine the predicates with !IsEndValueReached, so they now live
  next to the predicates they wrap.

- timestr.go — wall-clock time-string conversions. Three formats with
  different shapes (HH:MM[:SS] → seconds for baremin math, → decimal
  hours for `buzz add`, and "3:00 PM"/"15:00" → seconds-from-midnight
  for `buzz deadline`) now share one file. Co-located, not unified —
  each callsite needs different info.

- slope.go — bright-line geometry: rate-per-day from Roadall, with
  fallback to g.Rate when roadall isn't usable. The "how fast is the
  bright line moving right now" concept, separate from time.

No behaviour change. beeminder.go shrinks to types + sort + parse
helpers (421 → 250). The tomorrow-view baremin math
(bareminByEndOfTomorrowAt and helpers) stays in main.go as view-specific
orchestration — candidate for a future split alongside the goal-table
renderer (#248).

Refs #250
