+++ beeminder.go
@@ -1,10 +1,8 @@
 package main
 
 import (
-	"fmt"
 	"sort"
 	"strings"
-	"time"
 )
 
 // Goal represents a Beeminder goal with relevant fields
@@ -177,106 +175,6 @@ func ParseBareminValue(baremin string) string {
 	return value
 }
 
-// IsDueToday checks if a goal is due today (on or before midnight tonight)
-func IsDueToday(losedate int64) bool {
-	return IsDueTodayAt(losedate, time.Now())
-}
-
-// IsDueTodayAt checks if a goal is due today relative to a given time
-func IsDueTodayAt(losedate int64, now time.Time) bool {
-	goalTime := time.Unix(losedate, 0)
-
-	// Get start of tomorrow (midnight tonight)
-	startOfTomorrow := time.Date(now.Year(), now.Month(), now.Day()+1, 0, 0, 0, 0, now.Location())
-
-	// Goal is due today if it's due before the start of tomorrow
-	// This includes overdue goals and goals due later today
-	return goalTime.Before(startOfTomorrow)
-}
-
-// IsDueTomorrow checks if a goal is due tomorrow (between midnight tonight and midnight tomorrow)
-func IsDueTomorrow(losedate int64) bool {
-	return IsDueTomorrowAt(losedate, time.Now())
-}
-
-// IsDueTomorrowAt checks if a goal is due tomorrow relative to a given time
-func IsDueTomorrowAt(losedate int64, now time.Time) bool {
-	goalTime := time.Unix(losedate, 0)
-
-	// Get start of tomorrow (midnight tonight)
-	startOfTomorrow := time.Date(now.Year(), now.Month(), now.Day()+1, 0, 0, 0, 0, now.Location())
-	// Get start of day after tomorrow
-	startOfDayAfterTomorrow := time.Date(now.Year(), now.Month(), now.Day()+2, 0, 0, 0, 0, now.Location())
-
-	// Goal is due tomorrow if it's on or after midnight tonight but before the day after tomorrow
-	return !goalTime.Before(startOfTomorrow) && goalTime.Before(startOfDayAfterTomorrow)
-}
-
-// ParseDuration parses a duration string (e.g., "10m", "1h", "5d", "1w") and returns time.Duration
-// Supported formats: Nm (minutes), Nh (hours), Nd (days), Nw (weeks) where N is a number
-// Returns the duration and true on success, 0 and false on error
-func ParseDuration(durationStr string) (time.Duration, bool) {
-	if len(durationStr) < 2 {
-		// Need at least one character for number and one for unit
-		return 0, false
-	}
-
-	// Get the unit (last character)
-	unit := durationStr[len(durationStr)-1]
-
-	// Get the numeric part
-	numStr := durationStr[:len(durationStr)-1]
-
-	// Parse the number
-	var num float64
-	if _, err := fmt.Sscanf(numStr, "%f", &num); err != nil {
-		return 0, false
-	}
-
-	// Reject negative durations, which don't make sense for "due within" semantics
-	if num < 0 {
-		return 0, false
-	}
-
-	// Convert to duration based on unit
-	var duration time.Duration
-	switch unit {
-	case 'm', 'M':
-		duration = time.Duration(num * float64(time.Minute))
-	case 'h', 'H':
-		duration = time.Duration(num * float64(time.Hour))
-	case 'd', 'D':
-		duration = time.Duration(num * 24 * float64(time.Hour))
-	case 'w', 'W':
-		duration = time.Duration(num * 7 * 24 * float64(time.Hour))
-	default:
-		return 0, false
-	}
-
-	// Check for overflow: time.Duration is int64 nanoseconds
-	// Maximum duration is ~290 years (math.MaxInt64 nanoseconds)
-	// If the result is negative, it overflowed
-	if duration < 0 {
-		return 0, false
-	}
-
-	return duration, true
-}
-
-// IsDueWithin checks if a goal is due within the specified duration from now
-func IsDueWithin(losedate int64, duration time.Duration) bool {
-	return IsDueWithinAt(losedate, duration, time.Now())
-}
-
-// IsDueWithinAt checks if a goal is due within the specified duration from the given time
-func IsDueWithinAt(losedate int64, duration time.Duration, now time.Time) bool {
-	goalTime := time.Unix(losedate, 0)
-	cutoffTime := now.Add(duration)
-
-	// Goal is due within the duration if it's not after the cutoff time
-	return !goalTime.After(cutoffTime)
-}
-
 // IsEndValueReached reports whether the goal's current value has already met or passed
 // its end value (goalval). When this is true the bright red line has plateaued and the
 // goal effectively has no remaining work, so it shouldn't be surfaced as "due".
@@ -350,72 +248,3 @@ func IsDoLessGoal(goal Goal) bool {
 	return false
 }
 
-// FormatDueDate formats the losedate timestamp into a readable string
-func FormatDueDate(losedate int64) string {
-	return FormatDueDateAt(losedate, time.Now())
-}
-
-// FormatDueDateAt formats the losedate timestamp relative to a given time
-func FormatDueDateAt(losedate int64, now time.Time) string {
-	t := time.Unix(losedate, 0)
-
-	// Calculate duration until due
-	duration := t.Sub(now)
-	totalHours := duration.Hours()
-
-	if totalHours < 0 {
-		return "OVERDUE"
-	}
-
-	// If less than 1 day, show in hours or minutes
-	if totalHours < 24 {
-		if totalHours >= 1 {
-			// Show in hours (rounded down)
-			hours := int(totalHours)
-			return fmt.Sprintf("%dh", hours)
-		} else {
-			// Show in minutes (rounded down)
-			minutes := int(duration.Minutes())
-			if minutes < 1 {
-				return "0m"
-			}
-			return fmt.Sprintf("%dm", minutes)
-		}
-	}
-
-	// Show in days with "d" suffix
-	days := int(totalHours / 24)
-	return fmt.Sprintf("%dd", days)
-}
-
-// FormatAbsoluteDeadline formats the losedate timestamp as an absolute date/time string
-// Returns a compact format suitable for table display
-func FormatAbsoluteDeadline(losedate int64) string {
-	return FormatAbsoluteDeadlineAt(losedate, time.Now())
-}
-
-// FormatAbsoluteDeadlineAt formats the losedate timestamp as an absolute date/time string relative to a given time
-// Returns a compact format suitable for table display
-func FormatAbsoluteDeadlineAt(losedate int64, now time.Time) string {
-	// Convert Unix timestamp to the same timezone as now for accurate comparisons
-	t := time.Unix(losedate, 0).In(now.Location())
-
-	// Get start of today
-	startOfToday := time.Date(now.Year(), now.Month(), now.Day(), 0, 0, 0, 0, now.Location())
-	// Get start of tomorrow
-	startOfTomorrow := startOfToday.AddDate(0, 0, 1)
-
-	// If it's today, show time only (e.g., "3:04 PM")
-	if !t.Before(startOfToday) && t.Before(startOfTomorrow) {
-		return t.Format("3:04 PM")
-	}
-
-	// If it's tomorrow, show "tomorrow" + time (e.g., "tomorrow 3:04 PM")
-	startOfDayAfterTomorrow := startOfTomorrow.AddDate(0, 0, 1)
-	if !t.Before(startOfTomorrow) && t.Before(startOfDayAfterTomorrow) {
-		return "tomorrow " + t.Format("3:04 PM")
-	}
-
-	// For other dates, show date and time (e.g., "Jan 2 3:04 PM")
-	return t.Format("Jan 2 3:04 PM")
-}
