+++ main.go
@@ -504,29 +504,6 @@ func displayNextGoalWithTimestamp() {
 	fmt.Printf("\nRefreshing every %dm... (Press Ctrl+C to exit)\n", int(RefreshInterval.Minutes()))
 }
 
-// isDueTodayFilterAt returns true if the goal is due today (relative to now) and
-// hasn't already reached its end value. Exposed for deterministic time-based tests.
-func isDueTodayFilterAt(g Goal, now time.Time) bool {
-	return IsDueTodayAt(g.Losedate, now) && !IsEndValueReached(g)
-}
-
-// isDueTodayFilter returns true if the goal is due today and hasn't already reached its end value
-func isDueTodayFilter(g Goal) bool {
-	return isDueTodayFilterAt(g, time.Now())
-}
-
-// isDueTomorrowFilterAt returns true if the goal is due by the end of tomorrow
-// (i.e. due today or tomorrow) and hasn't already reached its end value. Goals
-// due today are included so the tomorrow view shows the full set of goals the
-// user must address to avoid a beemergency tomorrow. Exposed for deterministic
-// time-based tests.
-func isDueTomorrowFilterAt(g Goal, now time.Time) bool {
-	if IsEndValueReached(g) {
-		return false
-	}
-	return IsDueTodayAt(g.Losedate, now) || IsDueTomorrowAt(g.Losedate, now)
-}
-
 // bareminByEndOfTomorrowAt returns the baremin string to display for a goal in
 // the "due tomorrow" view. For goals already due tomorrow, the goal's existing
 // Baremin string is returned (it already reflects what's needed by tomorrow's
@@ -635,190 +612,6 @@ func tomorrowDaystampFor(g Goal, now time.Time) string {
 	shifted := now.Add(-time.Duration(g.Deadline) * time.Second).In(now.Location())
 	tomorrow := time.Date(shifted.Year(), shifted.Month(), shifted.Day(), 0, 0, 0, 0, now.Location()).AddDate(0, 0, 1)
 	return tomorrow.Format("20060102")
-}
-
-// parseTimeValue parses a "[-]HH:MM" or "[-]HH:MM:SS" string into total
-// seconds. The second return value reports whether the input included a
-// seconds component, so callers can preserve the original format on output.
-// Returns ok=false for anything that isn't exactly two or three colon-separated
-// non-negative integer fields with minutes and seconds < 60.
-func parseTimeValue(s string) (totalSeconds int, includeSeconds bool, ok bool) {
-	sign := 1
-	if strings.HasPrefix(s, "-") {
-		sign = -1
-		s = strings.TrimPrefix(s, "-")
-	}
-	parts := strings.Split(s, ":")
-	if len(parts) != 2 && len(parts) != 3 {
-		return 0, false, false
-	}
-	hh, err := strconv.Atoi(parts[0])
-	if err != nil {
-		return 0, false, false
-	}
-	mm, err := strconv.Atoi(parts[1])
-	if err != nil {
-		return 0, false, false
-	}
-	ss := 0
-	if len(parts) == 3 {
-		ss, err = strconv.Atoi(parts[2])
-		if err != nil {
-			return 0, false, false
-		}
-	}
-	// Reject malformed inputs (e.g. "1:75", "1:-05:00", "--1:30"). The leading
-	// sign has already been stripped, so any negative field — or minutes/
-	// seconds outside [0, 60) — means the string was malformed.
-	if hh < 0 || mm < 0 || mm >= 60 || ss < 0 || ss >= 60 {
-		return 0, false, false
-	}
-	return sign * (hh*3600 + mm*60 + ss), len(parts) == 3, true
-}
-
-// formatTimeValue formats a signed second count back into Beeminder's
-// colon-separated baremin style. When includeSeconds is true the output is
-// HH:MM:SS, otherwise HH:MM. Hours, minutes, and seconds are zero-padded and a
-// leading "+" is included for non-negative values. When dropping seconds, the
-// value is rounded to the nearest minute first so a fractional-minute bump
-// from the rate conversion doesn't silently undercount by up to 59 seconds.
-func formatTimeValue(totalSeconds int, includeSeconds bool) string {
-	if !includeSeconds {
-		totalSeconds = int(math.Round(float64(totalSeconds)/60.0)) * 60
-	}
-	sign := "+"
-	if totalSeconds < 0 {
-		sign = "-"
-		totalSeconds = -totalSeconds
-	}
-	h := totalSeconds / 3600
-	m := (totalSeconds % 3600) / 60
-	if includeSeconds {
-		return fmt.Sprintf("%s%02d:%02d:%02d", sign, h, m, totalSeconds%60)
-	}
-	return fmt.Sprintf("%s%02d:%02d", sign, h, m)
-}
-
-// isKnownRunits reports whether the given runits string is one of the values
-// Beeminder uses (y/m/w/d/h). Callers that need a dimensionally-correct
-// per-day conversion should bail out for anything else.
-func isKnownRunits(runits string) bool {
-	switch runits {
-	case "y", "m", "w", "d", "h":
-		return true
-	}
-	return false
-}
-
-// ratePerDay converts a rate expressed in the given runits into an equivalent
-// amount per day. Supports the runits Beeminder reports: y, m, w, d, h. For
-// unrecognised units the rate is returned unchanged.
-func ratePerDay(rate float64, runits string) float64 {
-	switch runits {
-	case "y":
-		return rate / 365.0
-	case "m":
-		return rate / 30.0
-	case "w":
-		return rate / 7.0
-	case "d":
-		return rate
-	case "h":
-		return rate * 24.0
-	default:
-		return rate
-	}
-}
-
-// slopePerDayAt returns the slope of the goal's bright line at the given
-// moment, expressed in the goal's gunits per day.
-//
-// For goals with a piecewise schedule (rate changes scheduled in the future),
-// g.Rate reports the rate at the *end* of the graph rather than the rate of
-// the segment the user is currently on — so a "1 h/day now, dropping to
-// 0.1 h/day later" goal would otherwise be bumped by the wrong amount. Resolve
-// the current segment from g.Roadall and use its slope; fall back to g.Rate
-// only when roadall isn't usable.
-func slopePerDayAt(g Goal, t time.Time) (float64, bool) {
-	if slope, ok := roadallSlopePerDayAt(g, t); ok {
-		return slope, true
-	}
-	if g.Rate == nil {
-		return 0, false
-	}
-	if !isKnownRunits(g.Runits) {
-		return 0, false
-	}
-	return ratePerDay(*g.Rate, g.Runits), true
-}
-
-// roadallSlopePerDayAt walks Beeminder's piecewise bright line (roadall) and
-// returns the slope of the segment containing t, in gunits/day. Rows are
-// [t, v, r] triples with exactly one of v/r null per row past the first
-// anchor row. When the row's rate is given, it's converted via ratePerDay;
-// when only values are given, the slope is computed directly from
-// (Δvalue / Δtime) so the goal's runits don't matter.
-//
-// Returns ok=false for a missing/short roadall, unparseable rows, or when t
-// is past the goal's end date (caller falls back to g.Rate in that case).
-func roadallSlopePerDayAt(g Goal, t time.Time) (float64, bool) {
-	if len(g.Roadall) < 2 {
-		return 0, false
-	}
-	target := float64(t.Unix())
-	for i := 1; i < len(g.Roadall); i++ {
-		cur := g.Roadall[i]
-		// A malformed boundary row makes the road ambiguous; fail fast so
-		// the caller falls back to g.Rate rather than silently advancing
-		// to a later segment that doesn't actually contain `target`.
-		if len(cur) < 3 || cur[0] == nil {
-			return 0, false
-		}
-		// Per Beeminder spec, non-anchor rows must have exactly one of
-		// value/rate set. Both nil or both set means the row is ambiguous.
-		if (cur[1] == nil) == (cur[2] == nil) {
-			return 0, false
-		}
-		prev := g.Roadall[i-1]
-		if len(prev) < 3 || prev[0] == nil {
-			return 0, false
-		}
-		// Validate prev's shape: the first row is the start anchor (value
-		// set, rate nil); subsequent rows follow the same one-of-v-or-r
-		// constraint as cur.
-		if i == 1 {
-			if prev[1] == nil || prev[2] != nil {
-				return 0, false
-			}
-		} else if (prev[1] == nil) == (prev[2] == nil) {
-			return 0, false
-		}
-		// `target` is before the road even starts — don't fall through to
-		// segment 1 just because target <= cur[0].
-		if target < *prev[0] {
-			return 0, false
-		}
-		if target > *cur[0] {
-			continue
-		}
-		// `t` falls in the segment ending at this row.
-		if cur[2] != nil {
-			if !isKnownRunits(g.Runits) {
-				return 0, false
-			}
-			return ratePerDay(*cur[2], g.Runits), true
-		}
-		// Rate not specified — derive from (Δvalue / Δtime).
-		if prev[1] == nil || cur[1] == nil {
-			return 0, false
-		}
-		seconds := *cur[0] - *prev[0]
-		if seconds <= 0 {
-			return 0, false
-		}
-		return (*cur[1] - *prev[1]) / seconds * 86400.0, true
-	}
-	return 0, false
 }
 
 // isDoLessFilter returns true if the goal is a do-less type goal
@@ -993,14 +786,6 @@ func losedateByEndOfTomorrowAt(g Goal, now time.Time) int64 {
 		return g.Losedate
 	}
 	return time.Unix(g.Losedate, 0).In(now.Location()).AddDate(0, 0, 1).Unix()
-}
-
-// dueLaterTodayAt reports whether a goal's losedate falls between now and the
-// start of tomorrow. Used as the gating predicate for the tomorrow-view
-// bumping helpers: overdue goals (losedate < now) and goals due tomorrow or
-// later are left untouched so their actual deadline keeps showing.
-func dueLaterTodayAt(g Goal, now time.Time) bool {
-	return g.Losedate >= now.Unix() && IsDueTodayAt(g.Losedate, now)
 }
 
 // handleLessCommand outputs all do-less type goals
@@ -1594,45 +1379,6 @@ func handleChargeCommand() {
 
 	// Check for updates and display message if available
 	fmt.Print(getUpdateMessage())
-}
-
-// parseTimeToDeadlineOffset parses a time string (e.g., "3:00 PM", "15:00") into
-// a deadline offset in seconds from midnight, as used by the Beeminder API.
-func parseTimeToDeadlineOffset(timeStr string) (int, error) {
-	trimmed := strings.TrimSpace(timeStr)
-
-	// Try 12-hour format first (e.g., "3:00 PM", "11:30 AM")
-	t, err := time.Parse("3:04 PM", strings.ToUpper(trimmed))
-	if err != nil {
-		// Try 24-hour format (e.g., "15:00", "23:30")
-		t, err = time.Parse("15:04", trimmed)
-	}
-	if err != nil {
-		return 0, fmt.Errorf("invalid time format %q (expected e.g. \"3:00 PM\" or \"15:00\")", timeStr)
-	}
-
-	hour := t.Hour()
-	minute := t.Minute()
-
-	// Beeminder does not allow deadlines between 6:01 AM and 6:59 AM.
-	// Allowed ranges: 12:00 AM–6:00 AM (nightowl) and 7:00 AM–11:59 PM (earlybird).
-	// https://help.beeminder.com/article/14-deadline#allowed
-	if hour == 6 && minute > 0 {
-		return 0, fmt.Errorf("deadlines between 6:01 AM and 6:59 AM are not allowed by Beeminder")
-	}
-
-	offset := hour*3600 + minute*60
-
-	// Beeminder deadline offsets use seconds from midnight.
-	// Range: -61200 (7:00 AM) to 21600 (6:00 AM)
-	// Times from 7:00-23:59 are negative offsets (before midnight)
-	// Times from 0:00-6:00 are positive offsets (after midnight)
-	// https://forum.beeminder.com/t/api-deadline/10666
-	if hour > 6 {
-		offset = offset - 24*3600
-	}
-
-	return offset, nil
 }
 
 func handleDeadlineCommand() {
